--- src/client/components/CollaboratorAccessModal/CollaboratorAccessModalBody/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CollaboratorAccessModalBody'
